+++ src/Collection/AuditedCollection.php
@@ -80,7 +80,9 @@ class AuditedCollection implements Collection
     }
 
     /**
-     * @return true
+     * @return void
+     *
+     * @phpstan-ignore-next-line https://github.com/phpstan/phpstan-doctrine/pull/560
      */
     #[\ReturnTypeWillChange]
     public function add(mixed $element)
@@ -293,9 +295,10 @@ class AuditedCollection implements Collection
     /**
      * @return Collection<TKey, T>
      *
-     * @phpstan-param \Closure(T, int|string):bool $p
-     * @psalm-param \Closure(T=):bool $p
-     * @psalm-return Collection<TKey, T>
+     * @phpstan-param \Closure(T, TKey):bool $p
+     * @phpstan-return Collection<TKey, T>
+     *
+     * @psalm-suppress MoreSpecificImplementedParamType https://github.com/doctrine/collections/pull/411
      */
     #[\ReturnTypeWillChange]
     public function filter(\Closure $p)
@@ -323,9 +326,7 @@ class AuditedCollection implements Collection
      *
      * @phpstan-template U
      * @phpstan-param \Closure(T):U $func
-     * @psalm-return Collection<TKey, U>
-     *
-     * @psalm-suppress ImplementedParamTypeMismatch,InvalidArgument
+     * @phpstan-return Collection<TKey, U>
      */
     #[\ReturnTypeWillChange]
     public function map(\Closure $func)
@@ -339,7 +340,9 @@ class AuditedCollection implements Collection
      * @return array<Collection<TKey, T>>
      *
      * @phpstan-param \Closure(TKey, T):bool $p
-     * @psalm-return array{0: Collection<TKey, T>, 1: Collection<TKey, T>}
+     * @phpstan-return array{0: Collection<TKey, T>, 1: Collection<TKey, T>}
+     *
+     * @psalm-suppress MoreSpecificImplementedParamType https://github.com/doctrine/collections/pull/411
      */
     #[\ReturnTypeWillChange]
     public function partition(\Closure $p)
@@ -454,7 +457,6 @@ class AuditedCollection implements Collection
      * @return T|null
      *
      * @phpstan-return T|null
-     * @psalm-return mixed|null
      */
     #[\ReturnTypeWillChange]
     public function findFirst(\Closure $p)